Multimodal-Human-Intention-Detection-for-Upper-Limb-Exoskeleton-Assistance-in-Construction-Work-main/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
pandas>=1.5.0
opencv-python>=4.7.0
keyboard>=0.13.5
ultralytics>=8.0.0
pupil-labs-realtime-api>=0.1.13
ximu3>=1.0.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.2
scikit-learn>=1.3.0
